refactor(phase-48.1): split interface_writer.py from gromacs_writer (no behavior change)
- Move write_interface_gro_file + write_interface_top_file to interface_writer.py (~549 lines, byte-verbatim move verified via AST source-segment comparison)
- gromacs_writer.py re-exports from interface_writer (49 symbols preserved; object identity verified gromacs_writer.write_interface_gro_file is interface_writer.write_interface_gro_file = True)
- PRESERVED 'if custom_guest_info:' gate in write_interface_gro_file (NOT unified with custom_active — research §3 'custom_active opt-in gate' divergence + §7 risk #3); write_interface_top_file keeps the standard 'custom_active = (...)' pattern — both divergences preserved byte-identically
- Trimmed import list to exactly what the two functions use (InterfaceStructure + count_guest_atoms + 11 _shared symbols + 6 _gro_format helpers); no unused imports
- gromacs_writer.py: 2902 -> 2401 lines (net -501); interface_writer.py: 549 lines (exceeds min_lines 380)
- 14 byte-equivalence + 27 interface/custom-guest grompp + 15 tip4p_ice_lj_values (scan-all-files includes new module) + 127 tests/test_output regression ALL PASS; 1840 total tests (no loss)
